requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12
numpy>=1.26.4
openai-whisper
audioop-lts

# Voice Activity Detection
silero-vad==5.1
torch>=2.0.0
torchaudio>=2.0.0

# Local LLM for Grammar Correction
ollama